requirements.txt
# ==============================================
# FLOORBALL VISION - Python Dependencies
# ==============================================
# Installation: pip install -r requirements.txt
# ==============================================

# --------------------------------------------
# WEB FRAMEWORK
# --------------------------------------------
flask>=3.0.0
flask-sqlalchemy>=3.1.0
flask-migrate>=4.0.0

# --------------------------------------------
# DATENBANK
# --------------------------------------------
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
alembic>=1.13.0

# --------------------------------------------
# TASK QUEUE
# --------------------------------------------
celery[redis]>=5.3.0
redis>=5.0.0

# --------------------------------------------
# MACHINE LEARNING / YOLO
# --------------------------------------------
ultralytics>=8.3.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
numpy>=1.24.0
supervision>=0.25.0

# --------------------------------------------
# LABEL STUDIO
# --------------------------------------------
label-studio-sdk>=0.0.32

# --------------------------------------------
# VIDEO / MEDIA
# --------------------------------------------
yt-dlp>=2024.1.0
Pillow>=10.0.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.8

# --------------------------------------------
# UTILITIES
# --------------------------------------------
python-dotenv>=1.0.0
pyyaml>=6.0.0
requests>=2.31.0
tqdm>=4.66.0
click>=8.1.0

# --------------------------------------------
# DATA PROCESSING
# --------------------------------------------
pandas>=2.0.0
scipy>=1.11.0

# --------------------------------------------
# DEVELOPMENT
# --------------------------------------------
pytest>=7.4.0
ipykernel>=6.0.0
